+++ app/src/main/java/com/app/arcabyolimpo/presentation/navigation/NavGraph.kt
@@ -18,9 +18,11 @@ import com.app.arcabyolimpo.data.remote.interceptor.SessionManager
 import com.app.arcabyolimpo.domain.model.auth.UserRole
 import com.app.arcabyolimpo.presentation.common.components.LoadingShimmer
 import com.app.arcabyolimpo.presentation.screens.accountactivation.AccountActivationScreen
+import com.app.arcabyolimpo.presentation.screens.attendance.AttendanceListScreen
 import com.app.arcabyolimpo.presentation.screens.beneficiary.AddNewBeneficiaryScreen
 import com.app.arcabyolimpo.presentation.screens.beneficiary.BeneficiaryDetailScreen
 import com.app.arcabyolimpo.presentation.screens.beneficiary.BeneficiaryListScreen
+import com.app.arcabyolimpo.presentation.screens.beneficiary.ModifyBeneficiaryScreen
 import com.app.arcabyolimpo.presentation.screens.capacitations.DisabilitiesList
 import com.app.arcabyolimpo.presentation.screens.capacitations.DisabilitiesListScreen
 import com.app.arcabyolimpo.presentation.screens.home.assistant.CollaboratorHomeScreen
@@ -31,22 +33,24 @@ import com.app.arcabyolimpo.presentation.screens.passwordrecovery.PasswordRecove
 import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordRegistrationScreen
 import com.app.arcabyolimpo.presentation.screens.passwordregisteration.PasswordRegistrationSuccessScreen
 import com.app.arcabyolimpo.presentation.screens.product.addProduct.ProductAddScreen
-import com.app.arcabyolimpo.presentation.screens.product.updateProduct.ProductUpdateScreen
 import com.app.arcabyolimpo.presentation.screens.product.list.ProductListScreen
 import com.app.arcabyolimpo.presentation.screens.product.productDetail.ProductDetailScreen
+import com.app.arcabyolimpo.presentation.screens.product.updateProduct.ProductUpdateScreen
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchDetail.ProductBatchDetailScreen
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchModify.ProductBatchModifyScreen
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchRegister.ProductBatchRegisterScreen
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchesList.ProductBatchesListScreen
 import com.app.arcabyolimpo.presentation.screens.qr.qr.QrScreen
+import com.app.arcabyolimpo.presentation.screens.qr.scanqr.ScanQrScreen
+import com.app.arcabyolimpo.presentation.screens.qr.scanresult.ScanResultScreen
 import com.app.arcabyolimpo.presentation.screens.qr.workshopselection.QrWorkshopsListScreen
 import com.app.arcabyolimpo.presentation.screens.splash.SplashScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyAdd.SupplyAddScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyBatchList.SupplyBatchListScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyDetail.SuppliesDetailScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyList.SupplyListScreen
-import com.app.arcabyolimpo.presentation.screens.supply.supplybatchmodify.SupplyBatchModifyScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplyUpdate.SupplyUpdateScreen
+import com.app.arcabyolimpo.presentation.screens.supply.supplybatchmodify.SupplyBatchModifyScreen
 import com.app.arcabyolimpo.presentation.screens.supply.supplybatchregister.SupplyBatchRegisterScreen
 import com.app.arcabyolimpo.presentation.screens.tokenverification.TokenVerificationFailedScreen
 import com.app.arcabyolimpo.presentation.screens.tokenverification.TokenVerificationViewModel
@@ -86,6 +90,10 @@ sealed class Screen(
         fun createRoute(userId: String) = "update_user/$userId"
     }
 
+    object AttendanceList : Screen("attendance_list/{userId}") {
+        fun createRoute(userId: String) = "attendance_list/$userId"
+    }
+
     object TokenVerification : Screen("user/verify-token?token={token}") {
         fun createRoute(token: String) = "user/verify-token?token=$token"
     }
@@ -114,7 +122,7 @@ sealed class Screen(
 
     object AddNewWorkshop : Screen("workshop/add")
 
-    object ModifyWorkshop : Screen("workshop/modify/{idTaller}"){
+    object ModifyWorkshop : Screen("workshop/modify/{idTaller}") {
         fun createRoute(idTaller: String) = "workshop/modify/$idTaller"
     }
 
@@ -129,6 +137,10 @@ sealed class Screen(
     }
 
     object AddNewBeneficiary : Screen("beneficiary/create")
+
+    object ModifyBeneficiary : Screen("beneficiary/update/{beneficiaryId}"){
+        fun createRoute(beneficiaryId: String) = "beneficiary/update/$beneficiaryId"
+    }
 
     object CapacitationScreen : Screen("/disabilities/list")
 
@@ -166,7 +178,7 @@ sealed class Screen(
             idInsumo: String,
         ) = "supply_batch_list/dates/$date?idInsumo=$idInsumo"
     }
-    
+
     object ProductDetail : Screen("product/{productId}") {
         fun createRoute(productId: String) = "product/$productId"
     }
@@ -189,6 +201,10 @@ sealed class Screen(
             workshopName: String,
         ) = "qr/workshop_selection/show_qr/$workshopID/$workshopName"
     }
+
+    object ScanQr : Screen("qr/scan_qr")
+
+    object ValidateQr : Screen("qr/scan_qr/validate_qr")
 }
 
 /**
@@ -413,7 +429,7 @@ fun ArcaNavGraph(
 
         /** Collaborator Home Screen */
         composable(Screen.CollaboratorHome.route) {
-            CollaboratorHomeScreen()
+            CollaboratorHomeScreen(navController)
         }
 
         /** Scholar Home Screen */
@@ -435,6 +451,9 @@ fun ArcaNavGraph(
                     navController.navigate(Screen.UpdateUserScreen.createRoute(id))
                 },
                 onDeleteClick = { navController.popBackStack() },
+                onAttendanceClick = { id ->
+                    navController.navigate(Screen.AttendanceList.createRoute(id))
+                },
             )
         }
 
@@ -463,9 +482,20 @@ fun ArcaNavGraph(
                         ?.savedStateHandle
                         ?.set("refresh_detail_$userId", true)
                     navController.popBackStack()
-                }
+                },
             )
         }
+
+        /** Attendance List Screen */
+        composable(
+            route = Screen.AttendanceList.route,
+            arguments = listOf(navArgument("userId") { type = NavType.StringType }),
+        ) {
+            AttendanceListScreen(
+                onBackClick = { navController.popBackStack() },
+            )
+        }
+
 
         /**
          * Workshops List Screen.
@@ -503,11 +533,13 @@ fun ArcaNavGraph(
             arguments = listOf(navArgument("id") { type = NavType.StringType }),
         ) { backStackEntry ->
             val workshopId = backStackEntry.arguments?.getString("id") ?: ""
-            WorkshopDetailScreen(navController,
+            WorkshopDetailScreen(
+                navController,
                 workshopId,
                 onModifyClick = { workshopId ->
                     navController.navigate(Screen.ModifyWorkshop.createRoute(workshopId))
-                })
+                },
+            )
         }
 
         /**
@@ -534,13 +566,12 @@ fun ArcaNavGraph(
         composable(
             route = Screen.ModifyWorkshop.route,
             arguments = listOf(navArgument("idTaller") { type = NavType.StringType }),
-
-        ){ backStackEntry ->
+        ) { backStackEntry ->
             val workshopId = backStackEntry.arguments?.getString("idTaller") ?: ""
             modifyWorkshopScreen(
                 navController = navController,
                 viewModel = hiltViewModel(),
-                workshopId = workshopId
+                workshopId = workshopId,
             )
         }
 
@@ -566,7 +597,7 @@ fun ArcaNavGraph(
                 },
                 onBackClick = {
                     navController.popBackStack()
-                }
+                },
             )
         }
 
@@ -689,8 +720,29 @@ fun ArcaNavGraph(
         ) {
             BeneficiaryDetailScreen(
                 onBackClick = { navController.popBackStack() },
-                onModifyClick = { /* TODO: Lógica de VM */ },
+                onModifyClick = { beneficiaryId ->
+                    navController.navigate(Screen.ModifyBeneficiary.createRoute(beneficiaryId))
+                },
                 viewModel = hiltViewModel(),
+                beneficiaryId = it.arguments?.getString("beneficiaryId") ?: ""
+            )
+        }
+
+        /**
+         * Modify Beneficiary Screen.
+         *
+         * Allows the modification of an existing and active beneficiary.
+         */
+        composable(
+            route = Screen.ModifyBeneficiary.route,
+            arguments = listOf(navArgument("beneficiaryId") { type = NavType.StringType }),
+
+            ){ backStackEntry ->
+            val beneficiaryId = backStackEntry.arguments?.getString("beneficiaryId") ?: ""
+            ModifyBeneficiaryScreen(
+                navController = navController,
+                viewModel = hiltViewModel(),
+                beneficiaryId = beneficiaryId
             )
         }
 
@@ -827,16 +879,15 @@ fun ArcaNavGraph(
                 onAddProductClick = {
                     navController.navigate(Screen.ProductAdd.route)
                 },
-                onBackClick ={
-                  navController.navigateUp()
+                onBackClick = {
+                    navController.navigateUp()
                 },
             )
         }
 
-
         composable(
             route = Screen.ProductDetail.route,
-            arguments = listOf(navArgument("productId") { type = NavType.StringType } ),
+            arguments = listOf(navArgument("productId") { type = NavType.StringType }),
         ) { backStackEntry ->
             val productId = backStackEntry.arguments?.getString("productId") ?: ""
             ProductDetailScreen(
@@ -851,13 +902,12 @@ fun ArcaNavGraph(
             )
         }
 
-
-
         composable(
             route = Screen.ProductUpdate.route,
-            arguments = listOf(
-                navArgument("idProduct") { type = NavType.StringType }
-            )
+            arguments =
+                listOf(
+                    navArgument("idProduct") { type = NavType.StringType },
+                ),
         ) {
             ProductUpdateScreen(
                 onModifyClick = {
@@ -887,7 +937,7 @@ fun ArcaNavGraph(
             val date = backStackEntry.arguments?.getString("date") ?: ""
             val idInsumo = backStackEntry.arguments?.getString("idInsumo") ?: ""
             android.util.Log.d("NavGraph", "SupplyBatchList args received: date=$date, idInsumo=$idInsumo")
-                SupplyBatchListScreen(
+            SupplyBatchListScreen(
                 supplyId = idInsumo,
                 supplyName = "",
                 date = date,
@@ -903,7 +953,7 @@ fun ArcaNavGraph(
 
         composable(
             route = Screen.SupplyUpdate.route,
-            arguments = listOf(navArgument("idSupply") { type = NavType.StringType })
+            arguments = listOf(navArgument("idSupply") { type = NavType.StringType }),
         ) { backStackEntry ->
 
             SupplyUpdateScreen(
@@ -946,6 +996,32 @@ fun ArcaNavGraph(
                 onBackClick = { navController.popBackStack() },
                 workshopId = id,
                 workshopName = name,
+            )
+        }
+
+        composable(Screen.ScanQr.route) {
+            ScanQrScreen(
+                onBackClick = { navController.popBackStack() },
+                onScanSuccess = { qrValue ->
+                    navController.currentBackStackEntry
+                        ?.savedStateHandle
+                        ?.set("qrValue", qrValue)
+                    navController.navigate(Screen.ValidateQr.route)
+                },
+            )
+        }
+
+        composable(route = Screen.ValidateQr.route) { backStackEntry ->
+            val qrValue =
+                navController.previousBackStackEntry
+                    ?.savedStateHandle
+                    ?.get<String>("qrValue")
+
+            ScanResultScreen(
+                onBackClick = {
+                    navController.popBackStack()
+                },
+                qrValue = qrValue,
             )
         }
     }