+++ CHANGELOG.md
@@ -1,7 +1,20 @@
+## [0.2.0] - 01/12/2025
+### Added
+- Modify beneficiary
+
+### Fixed
+- Debug modify beneficiary
+
 ## [0.2.0] - 30/11/2025
+### Added
+- feature for scanning qr code for attendance
 
 ### Changed
 - Added a wey to prevent certain chars and emojis in workshop inputs. 
+
+### Fixed
+- Fixed attendance date/time formatting.
+- Improved "View Attendance" button alignment and styling.
 
 ## [0.2.0] - 29/11/2025
 ### Added
@@ -19,14 +32,15 @@
 - Change beneficiary detail for showing more disabilities
 - Change so now Beneficiaries can be registered with more disabilities
 
+## [0.1.0] - 25/11/2025
+### Added
+- US for consult attendance
+
 ## [0.1.0] - 20/11/2025
 ### Added
 - ModifyProduct validation fields
 - RegisterProduct validation fields
 - Arrow to return register and modify product
-
-## [0.1.0] - 20/11/2025
-### Added
 - US for filter and order workshops.
 - Functional go back button to Supplies List Screen
 - Functional go back buttons to Orders and Inventory Screens